Made jumpBetweenOpeningAndClosingSymbols's behavior consistent for symbols with length > 1
This command is interested in the four positions of a symbol pair:
(1) The outside of a opening symbol
(2) The inside of a opening symbol
(3) The inside of a closing symbol
(4) The outside of a closing symbol

The behavior made in b3795cab feels weird (inconsistent) for the following situation:
* The highlighted symbols have length > 1
* The cursor is not at any of those positions

Before (b3795cab):
------------------
When this command is executed, if the cursor is not at any of those positions,
the cursor always jumps to (1).

Example:

  |#region           #region
   // ...       ->   // ...
   #endregion        #endregion|

   #|region         |#region
   // ...       ->   // ...
   #endregion        #endregion

   #regio|n         |#region
   // ...       ->   // ...
   #endregion        #endregion

   #region|          #region
   // ...       ->   // ...
   #endregion       |#endregion

   #region           #region|
   // ...       ->   // ...
  |#endregion        #endregion

   #region          |#region
   // ...       ->   // ...
   #|endregion       #endregion

   #region          |#region
   // ...       ->   // ...
   #end|region       #endregion

   #region          |#region
   // ...       ->   // ...
   #endregion|       #endregion

After (this commit):
--------------------
When this command is executed, if the cursor is not at any of those positions,
it calculates the distances towards those, and jumps to what it thinks is the "opposite" one.

Example:

  |#region           #region
   // ...       ->   // ...
   #endregion        #endregion|

   #|region          #region
   // ...       ->   // ...
   #endregion        #endregion|

   #regio|n          #region
   // ...       ->   // ...
   #endregion       |#endregion

   #region|          #region
   // ...       ->   // ...
   #endregion       |#endregion

   #region           #region|
   // ...       ->   // ...
  |#endregion        #endregion

   #region           #region|
   // ...       ->   // ...
   #|endregion       #endregion

   #region           #region|
   // ...       ->   // ...
   #end|region       #endregion

   #region          |#region
   // ...       ->   // ...
   #endregion|       #endregion

Whether or not this behavior is desirable in real coding experience is
to be questioned. At least it feels consistent now.